Fix: move switch account screen and translation to main src directory
