+++ web/src/components/SearchBox/SearchBox.pcss
