+++ web/src/components/style.pcss
@@ -48,3 +48,4 @@
 @import "./LockedComponent/LockedComponent.pcss";
 @import "./Beneficiaries/BeneficiariesCrypto/SelectBeneficiariesCrypto.pcss";
 @import "./WarningMessage/WarningMessage.pcss";
+@import "./SearchBox/SearchBox.pcss";
