+++ web/src/custom/components/index.tsx
@@ -4,4 +4,3 @@ export * from './HeaderProfile';
 export * from './CompanyAccount';
 export * from './DownloadBalance';
 export * from './CalendarPicker';
-export * from './SearchBox';
