+++ web/src/custom/components/style.pcss
@@ -4,5 +4,4 @@
 @import './CompanyAccount/CompanyAccount.pcss';
 @import './DownloadBalance/DownloadBalance.pcss';
 @import './CalendarPicker/CalendarPicker.pcss';
-@import './SearchBox/SearchBox.pcss';
 @import './WalletsHeader/WalletsHeader.pcss';
