--- web/src/custom/screens/index.tsx
@@ -1,1 +1,0 @@
-export * from './SwitchAccount';
