--- web/src/custom/screens/style.pcss
@@ -1,1 +1,0 @@
-@import './SwitchAccount/SwitchAccount.pcss';
