+++ web/src/custom/styles/style.pcss
@@ -1,7 +1,4 @@
 @import "./fonts.pcss";
 @import "./themes/style.pcss";
 @import "./colors.pcss";
-@import "../screens/style.pcss";
 @import "../components/style.pcss";
-
-/* @import "../mobile/styles/style.pcss"; */
