+++ web/src/screens/index.ts
@@ -22,3 +22,4 @@ export * from './SetupScreen';
 export * from './InternalTransfer';
 export * from './P2P';
 export * from './WalletsScreen';
+export * from './SwitchAccount';
