+++ web/src/screens/style.pcss
@@ -16,3 +16,4 @@
 @import "./SetupScreen/SetupScreen.pcss";
 @import "./WalletsScreen/WalletsScreen.pcss";
 @import "./P2P/style.pcss";
+@import "./SwitchAccount/SwitchAccount.pcss";
