+++ web/src/translations/en.ts
@@ -1349,6 +1349,21 @@ export const en = {
     'success.payment_method.updated': 'Payment Method: successfully updated',
     'success.p2p.dispute.submit': 'Dispute has been successfully submitted',
 
+    'page.body.profile.account.switch.search': 'Search by client name or UID',
+
+    'success.account.switch': 'Account has been switch successfully',
+
+    'page.body.dashboard.title': 'Account Dashboard',
+    'page.body.dashboard.asset': 'Asset',
+    'page.body.dashboard.totalLimitValue': 'Total limit value USD',
+    'page.body.dashboard.totalDailyLimit': 'Total daily limit',
+    'page.body.dashboard.available': 'Available',
+    'page.body.dashboard.used': 'Used',
+    'page.body.dashboard.trade': 'Trade',
+    'page.body.dashboard.tab.otc': 'OTC',
+
+    'page.body.wallets.tab.organization': 'Organization Overview',
+
     ...nationalitiesNames,
     ...customTranslations,
     ...mobileTranslations,
